+++ src/models/BibDetails.tsx
